Move light helper get_supported_color_modes (#51269)
